requirements.txt
langchain==0.0.267
openai==0.27.8
faiss-cpu==1.7.4
numpy>=1.24.3
tiktoken>=0.4.0
pytest>=7.3.1
python-dotenv>=1.0.0
async-timeout>=4.0.2
fastapi>=0.68.0
uvicorn>=0.15.0